Switch mock based tests of check_db_migrated to functional tests
The existing tests for check_db_migrated were hard to understand, as
they relied heavily on mocking out parts of the Pyramid and alembic
internals. They effectively tested the implementation of the
`check_db_migrated` function, rather than the outcomes. This change
introduces a new local fixture which downgrades the structure of
the test database for one test, allowing for tests to run in the
context of an outdated database.

The existing tests for `check_db_migrated` are then replaced with simple
invokations of `bootstrap()`, which causes `check_db_migrated` to be
called by the app startup (as `SKIP_CHECK_DB_MIGRATED` is not given).

The tests assert that the check is indeed called by `bootstrap` and
that `sys.exit` is either called or not, depending on which `db`
fixture is in use.