Enforce critical vars must be set
One of the features of bash's variable expansion system is to enforce that
the value of a variable must be defined using the `{VAR:?}` syntax.
This feature is supported by https://github.com/sayanarijit/expandvars,
used in this app to read the value of config vars from the environment.

Let's use this feature to enforce the same for the env vars in
`etc/production.ini`.